fix mkdir

+++ configure-many.py
@@ -235,8 +235,9 @@ if __name__ == "__main__":
         probability_addition_innode=0.1
         )
 
-    shutil.rmtree(os.path.normpath("./helix-1.0/configs"))
-    os.mkdir(os.path.normpath("./helix-1.0/configs"))
+    if os.path.isdir("./helix-1.0/configs"):
+        shutil.rmtree(os.path.normpath("./helix-1.0/configs"))
+        os.mkdir(os.path.normpath("./helix-1.0/configs"))
 
     if not os.path.isdir("./snapshots"):
         os.mkdir(os.path.normpath("./snapshots"))
